+++ lib/jido_flow/reduce.ex
@@ -182,6 +182,9 @@ defmodule Jido.Flow.Reduce do
           type: details.type
         })
 
+      :improper_list ->
+        Error.validation_error("#{owner} must be a proper list", %{path: path})
+
       :unsupported_expression ->
         Error.validation_error("#{owner} contains unsupported expression", %{
           path: path,